chore: improve IconButton, Image exports
